requirements.txt
requests
biopython
googletrans==4.0.0-rc1
